refactor(app): extract TraceTreeToolbar into its own file
Make the toolbar more compositional by moving it out of TraceTree.tsx
into a sibling TraceTreeToolbar.tsx.
